+++ infra/tf/dns_records.tf
@@ -4,7 +4,7 @@
 
 # Production domain (banterbus.games)
 resource "cloudflare_record" "prod_apex" {
-  zone_id = var.zone_id
+  zone_id = local.cloudflare_zone_id
   name    = "@"
   type    = "CNAME"
   content = local.tunnel_hostname # Dynamically generated from OpenBao
@@ -14,7 +14,7 @@ resource "cloudflare_record" "prod_apex" {
 
 # Production wildcard (*.banterbus.games)
 resource "cloudflare_record" "prod_wildcard" {
-  zone_id = var.zone_id
+  zone_id = local.cloudflare_zone_id
   name    = "*"
   type    = "CNAME"
   content = local.tunnel_hostname
@@ -24,7 +24,7 @@ resource "cloudflare_record" "prod_wildcard" {
 
 # Development domain (dev.banterbus.games)
 resource "cloudflare_record" "dev_apex" {
-  zone_id = var.zone_id
+  zone_id = local.cloudflare_zone_id
   name    = "dev"
   type    = "CNAME"
   content = local.tunnel_hostname # Same tunnel for dev
@@ -34,7 +34,7 @@ resource "cloudflare_record" "dev_apex" {
 
 # Development wildcard (*.dev.banterbus.games)
 resource "cloudflare_record" "dev_wildcard" {
-  zone_id = var.zone_id
+  zone_id = local.cloudflare_zone_id
   name    = "*.dev"
   type    = "A"
   content = "5.75.159.214"
@@ -44,7 +44,7 @@ resource "cloudflare_record" "dev_wildcard" {
 
 # ACME challenge records for SSL certificates (if using cert-manager)
 resource "cloudflare_record" "acme_challenge" {
-  zone_id = var.zone_id
+  zone_id = local.cloudflare_zone_id
   name    = "_acme-challenge"
   type    = "CNAME"
   content = local.tunnel_hostname
@@ -53,7 +53,7 @@ resource "cloudflare_record" "acme_challenge" {
 }
 
 resource "cloudflare_record" "acme_challenge_dev" {
-  zone_id = var.zone_id
+  zone_id = local.cloudflare_zone_id
   name    = "_acme-challenge.dev"
   type    = "CNAME"
   content = local.tunnel_hostname