+++ infra/tf/openbao.tf
@@ -58,14 +58,5 @@ EOT
 
 
 
-data "vault_kv_secret_v2" "cloudflare_tunnel" {
-  mount = "kv"
-  name  = "infra/cloudflare"
-}
-
-locals {
-  tunnel_name = data.vault_kv_secret_v2.cloudflare_tunnel.data["tunnel_name"]
-  tunnel_id   = data.vault_kv_secret_v2.cloudflare_tunnel.data["tunnel_id"]
-  tunnel_hostname = "${local.tunnel_id}.cfargotunnel.com"
-}
+# Cloudflare tunnel data and locals moved to secrets.tf